24) Botón de desactivar proveedor:
Muestra pantalla de error ya que el botón de eliminar Proveedor no funciona SQLSTATE[23000]: se produce, un error en una restricción de clave externa
('datasystem'.'compras', CONSTRAINT
'compras_proveedor_id_foreign'
	H- 9 desactivar un proveedor

Corrección:   se capturo el error de eliminar un proveedor con registro de compra de empleado (Violación de integridad referencial con compras), y se colocó un mensaje "No se puede eliminar el proveedor porque tiene compras asociadas"

app/Http/Controllers/ProveedorController.php

Línea 186-196

+++ app/Http/Controllers/ProveedorController.php
@@ -189,6 +189,12 @@ class ProveedorController extends Controller
      */
     public function destroy($id)
     {
+        $proveedor = Proveedor::find($id);
+
+        if ($proveedor->compra()->exists()) {
+            return redirect()->route('proveedor.index')->with("error", "No se puede eliminar el proveedor porque tiene compras asociadas.");
+        }
+
         Proveedor::destroy($id);
         return redirect()->route('proveedor.index')->with("error", "Se eliminó exitosamente el proveedor.");
     }
